memento-agent: extract factory functions
