+++ packages/memento-agent/index.ts
@@ -1,3 +1,4 @@
 // Path: packages/memento-agent/index.ts
 
+export * from './src/factory'
 export * from './src/mementoAgent'
